fix(components/{Sidebar.jsx,TopNavbar.jsx} pages/{index.jsx,admin/index.jsx}): rename `routers`.
