+++ packages/frontend/package.json
@@ -56,7 +56,7 @@
     "redux": "^4.0.1",
     "redux-saga": "^0.16.2",
     "resolve": "1.8.1",
-    "resource-oriented-request": "^0.1.5",
+    "resource-oriented-request": "^0.1.6",
     "sass-loader": "7.1.0",
     "style-loader": "0.23.0",
     "terser-webpack-plugin": "1.1.0",
