+++ packages/frontend/src/components/Sidebar.jsx
@@ -10,7 +10,7 @@ import {
   LinkContainer
 } from 'react-router-bootstrap'
 
-import routes from '@/routers/admin'
+import routes from '@/routes/admin'
 import Avatar from './Avatar'
 
 export default class Sidebar extends Component {
