+++ packages/frontend/src/network/login.js
@@ -3,3 +3,5 @@ import Base from './base'
 export default class Login extends Base {
   static pluralise = s => s
 }
+
+Login.plural(Login.model)
